Codechange: [Win32] Move Win32-specific font code to a seperate file.
