+++ src/CMakeLists.txt
@@ -155,6 +155,7 @@ add_files(
     fios_gui.cpp
     fontcache.cpp
     fontcache.h
+    fontcache_internal.h
     fontdetection.cpp
     fontdetection.h
     framerate_gui.cpp
