+++ src/os/windows/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_files(
     crashlog_win.cpp
+    font_win32.cpp
+    font_win32.h
     string_uniscribe.cpp
     string_uniscribe.h
     win32.cpp
